refactoring front end state management heavily to try and figure this out
