+++ src/App.js
@@ -1,10 +1,9 @@
 import React from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import PrivateRoute from "./components/PrivateRoute";
-import Dashboard from "./components/Dashboard";
+import PrivateRoute from "./utils/PrivateRoute";
+import Dashboard from "./components/Dashboard/Dashboard";
 import LandingPage from "./components/LandingPage";
 import Account from "./components/Account";
-import Lesson from "./components/Lesson";
 import QuizLandingPage from "./components/Quiz/QuizLandingPage.js";
 import Quiz from "./components/Quiz/Quiz";
 
@@ -16,21 +15,14 @@ function App() {
           <LandingPage />
         </Route>
         <Route exact path="/account/signup">
-          {/* passing a value so the proper tab is displayed 
-            without this it causes a memory leak */}
-          <Account value={0} />
+          <Account value={0} /> {/*value displays correct tab preventing memory leak*/}
         </Route>
         <Route exact path="/account/login">
-          {/* passing a value so the proper tab is displayed 
-            without this it causes a memory leak */}
-          <Account value={1} />
+          <Account value={1} /> {/*value displays correct tab preventing memory leak*/}
         </Route>
         <PrivateRoute>
           <Route path="/dashboard">
             <Dashboard />
-          </Route>
-          <Route exact path="/lesson">
-            <Lesson />
           </Route>
           <Route path="/quizLanding">
             <QuizLandingPage />
