--- src/actions/ResetArray.js
@@ -1,5 +1,0 @@
-export const RESET_SUCCESS = "RESET_SUCCESS";
-
-export const resetArray = () => (dispatch) => {
-  dispatch({ type: RESET_SUCCESS });
-};
