+++ src/components/Dashboard/DashboardCard.js
@@ -1,0 +1,18 @@
+import React from "react";
+import { useHistory } from "react-router-dom";
+import { connect } from "react-redux";
+
+const DashboardCard = (props) => {
+  const history = useHistory();
+  const purl = process.env.PUBLIC_URL;
+
+  return (
+    <>
+      <p>dashboard card</p>
+    </>
+  );
+};
+
+const mapDispatchToProps = {};
+
+export default connect(null, mapDispatchToProps)(DashboardCard);
