+++ src/index.js
@@ -5,15 +5,11 @@ import { createStore, applyMiddleware, combineReducers } from "redux";
 import thunk from "redux-thunk";
 import logger from "redux-logger";
 import { authReducer } from "./reducers/authReducer";
-import { levelsReducer } from "./reducers/levelsReducer";
-import { lessonReducer } from "./reducers/lessonReducer";
 import "./css/index.css"
 import App from "./App";
 
 const rootReducer = combineReducers({
-  user: authReducer,
-  level: levelsReducer,
-  lesson: lessonReducer,
+  authReducer,
 });
 
 const store = createStore(rootReducer, applyMiddleware(thunk, logger));
